makefile enhancements
git-svn-id: http://localhost/svn/entropybroker/trunk@233 6e73ccbf-f98f-4734-8601-04fa8acba097

+++ Makefile
@@ -152,25 +152,15 @@ plot: plot.o
 
 install: entropy_broker eb_server_audio eb_server_timers eb_server_v4l eb_server_stream eb_server_egd eb_client_linux_kernel eb_client_egd eb_test_egd_speed eb_server_linux_kernel eb_client_file eb_server_push_file eb_server_ext_proc eb_server_usb plot
 	mkdir -p $(BIN) $(ETC) $(VAR) $(PID) $(CACHE)
-	cp entropy_broker $(BIN)
-	cp eb_server_audio $(BIN)
-	cp eb_server_timers $(BIN)
-	cp eb_server_v4l $(BIN)
-	cp eb_server_stream $(BIN)
-	cp eb_server_egd $(BIN)
-	cp eb_client_linux_kernel $(BIN)
-	cp eb_client_egd $(BIN)
-	cp eb_test_egd_speed $(BIN)
-	cp eb_server_linux_kernel $(BIN)
-	cp eb_client_file $(BIN)
-	ln $(BIN)/eb_client_file $(BIN)/eb_client_kernel_generic
-	cp eb_server_push_file $(BIN)
-	cp eb_server_ext_proc $(BIN)
-	cp eb_server_usb $(BIN)
-	cp eb_server_ComScire_R2000KU $(BIN)
-	cp entropybroker.conf $(ETC)
-	cp password.txt $(ETC)
-	chmod 600 $(ETC)/password.txt
+	for file in entropy_broker eb_server_audio eb_server_timers eb_server_v4l eb_server_stream eb_server_egd eb_client_linux_kernel eb_client_egd eb_test_egd_speed eb_server_linux_kernel eb_client_file eb_server_push_file eb_server_ext_proc eb_server_usb eb_server_ComScire_R2000KU ; do \
+		test -e $(file) && cp $$file $(BIN) ; \
+	done
+	test -e $(BIN)/eb_client_file && \
+		(test -e $(BIN)/eb_client_kernel_generic || \
+		ln $(BIN)/eb_client_file $(BIN)/eb_client_kernel_generic)
+	test -e $(ETC)/entropybroker.conf || cp entropybroker.conf $(ETC)
+	test -e $(ETC)/entropybroker.conf && cp entropybroker.conf $(ETC)/entropybroker.conf.dist
+	test -e $(ETC)/password.txt || (cp password.txt $(ETC) ; chmod 600 $(ETC)/password.txt)
 
 clean:
 	rm -f $(OBJSeb) $(OBJSsa) $(OBJSst) $(OBJSsv) $(OBJSss)$(OBJSse) $(OBJSclk) $(OBJSte) $(OBJSsk) $(OBJScf) $(OBJSpf) $(OBJSep) $(OBJSsu) $(OBJScsr2000ku) $(OBJScle) $(OBJSse) plot.o entropy_broker core *.da *.gcov *.bb* eb_server_audio eb_server_timers eb_server_v4l eb_server_stream eb_server_egd eb_client_linux_kernel eb_client_egd eb_test_egd_speed eb_server_linux_kernel eb_client_file eb_server_push_file eb_server_ext_proc eb_server_usb eb_server_ComScire_R2000KU plot
